feature-#107 [피커] 휠 피커 개발
- 버그 (최초 아이템 선택 시 선택된 아이템 빨간색으로 표시 되야 함)
