+++ app/src/main/java/cookcook/nexters/com/amoogye/views/calc/presenter/CalcFragment.kt
@@ -1,57 +1,40 @@
 package cookcook.nexters.com.amoogye.views.calc.presenter
 
-import android.animation.AnimatorSet
-import android.animation.ValueAnimator
+import android.annotation.SuppressLint
 import android.content.Intent
 import android.graphics.Color
-import android.graphics.Typeface
+import android.os.Build
+import android.text.InputType
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.MotionEvent
 import android.view.View
 import android.view.ViewGroup
-import android.view.animation.AccelerateDecelerateInterpolator
 import android.widget.EditText
-import androidx.core.content.res.ResourcesCompat
+import android.widget.Toast
 import androidx.databinding.DataBindingUtil
+import androidx.lifecycle.Observer
 import com.baoyz.actionsheet.ActionSheet
 import cookcook.nexters.com.amoogye.R
 import cookcook.nexters.com.amoogye.base.BaseFragment
 import cookcook.nexters.com.amoogye.base.BaseNumberButton
-import cookcook.nexters.com.amoogye.base.BaseScrollPicker
+import cookcook.nexters.com.amoogye.base.wheel_picker.BaseWheelPicker
 import cookcook.nexters.com.amoogye.databinding.FragmentCalcBinding
-import cookcook.nexters.com.amoogye.views.calc.entity.EditTextType
-import cookcook.nexters.com.amoogye.views.calc.entity.UnitModel
-import cookcook.nexters.com.amoogye.utils.realmData
-import cookcook.nexters.com.amoogye.views.calc.history.CalcHistory
+import cookcook.nexters.com.amoogye.views.calc.entity.*
 import cookcook.nexters.com.amoogye.views.calc.history.CalcHistoryActivity
-import cookcook.nexters.com.amoogye.views.tools.*
-import cookcook.nexters.com.amoogye.views.tools.add_tools.MeasureUnitSaveData
-import io.realm.Realm
-import io.realm.RealmResults
+import cookcook.nexters.com.amoogye.views.calc.presenter.adapter.AdapterListener
 import kotlinx.android.synthetic.main.fragment_calc.*
-import kotlinx.android.synthetic.main.layout_calc_hide_edit_text.*
 import kotlinx.android.synthetic.main.layout_calc_visible_edit_text.*
 import kotlinx.android.synthetic.main.layout_unit_button_wrap.*
 import org.koin.android.ext.android.get
-import java.util.*
-import kotlin.collections.ArrayList
 
 
 class CalcFragment : BaseFragment() {
     override val layoutRes: Int = R.layout.fragment_calc
     override val isUseDataBinding = true
     lateinit var binding: FragmentCalcBinding
-    private val calculatorViewModel: CalculatorViewModel = get()
+    private val calculatorViewModel: CalculatorViewModel2 = get()
     private lateinit var unitRecyclerView: UnitButtonActivity
-
-    private val ingredientSelectStatus = arrayOf(true, false, true)
-    private val portionSelectStatus = arrayOf(false, true, true)
-    private val plusSelectStatus = arrayOf(false, false, true)
-
-    private var currentContainerCase: Int = 0
-
-    private lateinit var realm: Realm
 
     companion object {
         // 선택 선언 1 (Fragment를 싱글턴으로 사용 시)
@@ -66,51 +49,131 @@ class CalcFragment : BaseFragment() {
         }
     }
 
-    private val onClick: (number: String) -> Unit = {
-        calculatorViewModel.onNumberButtonClick(it)
+    private fun convertBooleanToVisibleType(isVisible: Boolean): Int {
+        return if (isVisible) {
+            View.VISIBLE
+        } else {
+            View.GONE
+        }
     }
 
-    private fun editTextWrapChange(type: EditTextType) {
+    private fun hideOrVisibleLayout(isVisibleHuman: Boolean = false, isVisibleMaterial: Boolean = false) {
+        layout_human_top_standard.visibility = convertBooleanToVisibleType(isVisibleHuman)
+        layout_human_bottom_standard.visibility = convertBooleanToVisibleType(isVisibleHuman)
+        layout_ingredient_bottom_standard.visibility = convertBooleanToVisibleType(isVisibleMaterial)
+    }
+
+    private val changeSelectedEditText = Observer<EditTextType> { type ->
         edit_twice_human_one.setBackgroundResource(R.drawable.number_input_non_focus_wrap_rounded_box)
         edit_twice_amount.setBackgroundResource(R.drawable.number_input_non_focus_wrap_rounded_box)
         edit_twice_unit.setBackgroundResource(R.drawable.number_input_non_focus_wrap_rounded_box)
         edit_twice_ingredient.setBackgroundResource(R.drawable.number_input_non_focus_wrap_rounded_box)
         edit_twice_human_two.setBackgroundResource(R.drawable.number_input_non_focus_wrap_rounded_box)
         edit_twice_tool.setBackgroundResource(R.drawable.number_input_non_focus_wrap_rounded_box)
-
-        when (type) {
-            EditTextType.HUMAN_ONE -> edit_twice_human_one.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
-            EditTextType.AMOUNT -> edit_twice_amount.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
-            EditTextType.UNIT -> edit_twice_unit.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
-            EditTextType.INGREDIENT -> edit_twice_ingredient.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
-            EditTextType.HUMAN_TWO -> edit_twice_human_two.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
-            EditTextType.TOOL -> edit_twice_tool.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
+        btn_calc_button.visibility = View.VISIBLE
+        layout_calc_result.visibility = View.GONE
+        calc_frame_layout.visibility = View.VISIBLE
+        when (type!!) {
+            EditTextType.HUMAN_ONE -> {
+                edit_twice_human_one.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
+                calculatorViewModel.calcKeyboardType.value = CalcLayoutState.NUMBER
+            }
+            EditTextType.AMOUNT -> {
+                edit_twice_amount.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
+                calculatorViewModel.calcKeyboardType.value = CalcLayoutState.NUMBER
+            }
+            EditTextType.UNIT -> {
+                edit_twice_unit.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
+                calculatorViewModel.calcKeyboardType.value = CalcLayoutState.UNIT
+                calculatorViewModel.selectedUnitType.value = CalcUnitType.NORMAL
+            }
+            EditTextType.INGREDIENT -> {
+                edit_twice_ingredient.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
+                calculatorViewModel.calcKeyboardType.value = CalcLayoutState.INGREDIENT
+            }
+            EditTextType.HUMAN_TWO -> {
+                edit_twice_human_two.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
+                calculatorViewModel.calcKeyboardType.value = CalcLayoutState.NUMBER
+            }
+            EditTextType.TOOL -> {
+                edit_twice_tool.setBackgroundResource(R.drawable.number_input_wrap_rounded_box)
+                calculatorViewModel.calcKeyboardType.value = CalcLayoutState.TOOL
+                calculatorViewModel.selectedUnitType.value = CalcUnitType.LIFE
+            }
+            EditTextType.NONE -> {
+                calculatorViewModel.calcKeyboardType.value = CalcLayoutState.NONE
+            }
         }
     }
 
-    private fun convertBottomButton(type: EditTextType) {
+    private val changeKeyboard = Observer<CalcLayoutState> { state ->
         calc_layout_button.visibility = View.GONE
         calc_layout_unit.visibility = View.GONE
         calc_layout_ingredient.visibility = View.GONE
 
-        when (type) {
-            EditTextType.HUMAN_ONE,
-            EditTextType.HUMAN_TWO,
-            EditTextType.AMOUNT -> calc_layout_button.visibility = View.VISIBLE
-
-            EditTextType.UNIT,
-            EditTextType.TOOL -> calc_layout_unit.visibility = View.VISIBLE
-
-            EditTextType.INGREDIENT -> calc_layout_ingredient.visibility = View.VISIBLE
+        when (state!!) {
+            CalcLayoutState.NUMBER -> {
+                calc_layout_button.visibility = View.VISIBLE
+            }
+            CalcLayoutState.INGREDIENT -> {
+                calc_layout_ingredient.visibility = View.VISIBLE
+            }
+            CalcLayoutState.TOOL,
+            CalcLayoutState.UNIT -> {
+                calc_layout_unit.visibility = View.VISIBLE
+            }
+            CalcLayoutState.NONE -> {
+                btn_calc_button.visibility = View.GONE
+                layout_calc_result.visibility = View.VISIBLE
+                calc_frame_layout.visibility = View.INVISIBLE
+            }
         }
     }
 
-    private fun editTextClickEvent(editText: EditText, type: EditTextType) {
+    private val changeUI = Observer<CalcTypeState> { state ->
+        when (state!!) {
+            CalcTypeState.MATERIAL -> {
+                txt_ingredient.setTextColor(Color.parseColor("#131c32"))
+                txt_human.setTextColor(Color.parseColor("#33131c32"))
+                txt_calc_plus.setTextColor(Color.parseColor("#33131c32"))
+                hideOrVisibleLayout(isVisibleHuman = false, isVisibleMaterial = true)
+                calculatorViewModel.selectedEditBox.value = EditTextType.AMOUNT
+            }
+            CalcTypeState.PERSONNEL -> {
+                txt_human.setTextColor(Color.parseColor("#131c32"))
+                txt_ingredient.setTextColor(Color.parseColor("#33131c32"))
+                txt_calc_plus.setTextColor(Color.parseColor("#33131c32"))
+                hideOrVisibleLayout(isVisibleHuman = true, isVisibleMaterial = false)
+                calculatorViewModel.selectedEditBox.value = EditTextType.HUMAN_ONE
+            }
+            CalcTypeState.MATERIAL_PERSONNEL -> {
+                txt_ingredient.setTextColor(Color.parseColor("#131c32"))
+                txt_human.setTextColor(Color.parseColor("#131c32"))
+                txt_calc_plus.setTextColor(Color.parseColor("#131c32"))
+                hideOrVisibleLayout(isVisibleHuman = true, isVisibleMaterial = true)
+                calculatorViewModel.selectedEditBox.value = EditTextType.HUMAN_ONE
+            }
+        }
+    }
+
+    private val visibleWeightEditText = Observer<Boolean> { isVisible ->
+        layout_ingredient_wrap.visibility = convertBooleanToVisibleType(isVisible)
+    }
+
+    private val onClick: (number: String) -> Unit = {
+        calculatorViewModel.onNumberButtonClick(it)
+    }
+
+    @SuppressLint("ClickableViewAccessibility")
+    private fun editTextSetting(editText: EditText, type: EditTextType) {
+        editText.inputType = InputType.TYPE_NULL
+        editText.setRawInputType(InputType.TYPE_CLASS_TEXT)
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            editText.showSoftInputOnFocus = false
+        }
         editText.setOnTouchListener { _, event ->
             if (event.action == MotionEvent.ACTION_DOWN) {
-                calculatorViewModel.setSelectedEditText(type)
-                editTextWrapChange(calculatorViewModel.getSelectedEditText()!!)
-                convertBottomButton(calculatorViewModel.getSelectedEditText()!!)
+                calculatorViewModel.onSelectedEditBox(type)
             }
 
             return@setOnTouchListener false
@@ -118,39 +181,45 @@ class CalcFragment : BaseFragment() {
     }
 
     override fun setupViews(view: View) {
+        calculatorViewModel.currentCalcState.observe(this, changeUI)
+        calculatorViewModel.useIngredient.observe(this, visibleWeightEditText)
+        calculatorViewModel.selectedEditBox.observe(this, changeSelectedEditText)
+        calculatorViewModel.calcKeyboardType.observe(this, changeKeyboard)
+
+        calculatorViewModel.selectedUnitType.observe(this, changeItemList)
+
+        BaseNumberButton(view, onClick)
+
+        editTextSetting(edit_twice_amount, EditTextType.AMOUNT)
+        editTextSetting(edit_twice_human_one, EditTextType.HUMAN_ONE)
+        editTextSetting(edit_twice_human_two, EditTextType.HUMAN_TWO)
+        editTextSetting(edit_twice_ingredient, EditTextType.INGREDIENT)
+        editTextSetting(edit_twice_tool, EditTextType.TOOL)
+        editTextSetting(edit_twice_unit, EditTextType.UNIT)
+
+        unitRecyclerView = UnitButtonActivity(view, adapterListener)
+        unitRecyclerView.addItems(calculatorViewModel.convertUnitItemList())
+
+        txt_unit_normal_sequence.setOnClickListener {
+            Toast.makeText(this.context, "구현 예정", Toast.LENGTH_SHORT).show()
+        }
+
+        txt_unit_recent_sequence.setOnClickListener {
+            Toast.makeText(this.context, "구현 예정", Toast.LENGTH_SHORT).show()
+        }
+
         btn_history.setOnClickListener {
             val intent = Intent(context, CalcHistoryActivity::class.java)
             startActivity(intent)
 
         }
-        btn_tip.setOnClickListener { calculatorViewModel.gazuaa("tool_tip 구현 예정") }
 
-        // edittext setting
-        calculatorViewModel.calculatorEditTextSetting(edit_twice_human_one)
-        editTextClickEvent(edit_twice_human_one, EditTextType.HUMAN_ONE)
-        calculatorViewModel.calculatorEditTextSetting(edit_twice_amount)
-        editTextClickEvent(edit_twice_amount, EditTextType.AMOUNT)
-        calculatorViewModel.calculatorEditTextSetting(edit_twice_unit)
-        editTextClickEvent(edit_twice_unit, EditTextType.UNIT)
-        calculatorViewModel.calculatorEditTextSetting(edit_twice_ingredient)
-        editTextClickEvent(edit_twice_ingredient, EditTextType.INGREDIENT)
-        calculatorViewModel.calculatorEditTextSetting(edit_twice_human_two)
-        editTextClickEvent(edit_twice_human_two, EditTextType.HUMAN_TWO)
-        calculatorViewModel.calculatorEditTextSetting(edit_twice_tool)
-        editTextClickEvent(edit_twice_tool, EditTextType.TOOL)
+        btn_tip.setOnClickListener { Toast.makeText(this.context, "구현 예정", Toast.LENGTH_SHORT).show() }
 
-        BaseNumberButton(view, onClick)
-
-        txt_ingredient.setOnClickListener {
-            changeCalcContainerLayout(1)
-        }
-
-        txt_human.setOnClickListener {
-            changeCalcContainerLayout(2)
-        }
-
-        btn_unit_changer.setOnClickListener {
-            callUnitSelector()
+        btn_calc_button.setOnClickListener {
+            if (calculatorViewModel.checkCalculable()) {
+                calculation()
+            }
         }
 
         btn_unit_changer_previous.setOnClickListener {
@@ -158,7 +227,11 @@ class CalcFragment : BaseFragment() {
                 return@setOnClickListener
             }
             calculatorViewModel.reduceIndex()
-            unitRecyclerView.addItems(selectUnitItems(calculatorViewModel.currentSelectedType))
+            unitRecyclerView.addItems(calculatorViewModel.convertUnitItemList())
+        }
+
+        btn_unit_changer.setOnClickListener {
+            callUnitSelector()
         }
 
         btn_unit_changer_next.setOnClickListener {
@@ -168,331 +241,47 @@ class CalcFragment : BaseFragment() {
             }
 
             calculatorViewModel.increaseIndex()
-            unitRecyclerView.addItems(selectUnitItems(calculatorViewModel.currentSelectedType))
-        }
-
-        val data = realm.where(MeasureUnit::class.java).equalTo("unitType", TYPE_FOOD).findAll()
-
-
-        val items: ArrayList<String> = arrayListOf()
-
-        data.map { items.add(it.unitNameBold) }
-
-        val picker = BaseScrollPicker(view, items)
-
-        picker.wheelView.setLoopListener {
-            calculatorViewModel.setIngredient(items.get(it))
-            calculatorViewModel.ingredientObject = data.find { j -> j.unitNameBold == items[it] }
-        }
-
-        itemChange(calculatorViewModel.flag - 1)
-//        unitRecyclerView = UnitButtonActivity(view)
-        unitRecyclerView.addItems(selectUnitItems(calculatorViewModel.currentSelectedType))
-
-        val initData = realm.where(MeasureUnit::class.java).equalTo("unitId", 5).findFirst()
-        calculatorViewModel.unitObject = UnitModel(
-            initData.unitNameBold,
-            initData.unitNameSoft,
-            initData.unitType,
-            initData.isWeight,
-            initData.unitValue
-        )
-        calculatorViewModel.setUnit(initData.unitNameBold)
-
-        btn_calc_button.setOnClickListener {
-            var humanOne = calculatorViewModel.humanOne.value
-            var humanOneValue = 1
-            if (!humanOne!!.isEmpty()) {
-                humanOneValue = humanOne.toInt()
-            }
-
-            var humanTwo = calculatorViewModel.humanTwo.value
-            var humanTwoValue = 1
-            if (!humanOne.isEmpty()) {
-                humanTwoValue = humanTwo!!.toInt()
-            }
-
-            var amount = calculatorViewModel.amount.value
-            var amoutValue = 1
-            if (!amount!!.isEmpty()) {
-                amoutValue = amount.toInt()
-            }
-
-            var unit = calculatorViewModel.unitObject
-            var tool = calculatorViewModel.toolObject
-            var ingredient = calculatorViewModel.ingredientObject
-
-            if (unit!!.abbreviation.isEmpty() || tool!!.abbreviation.isEmpty()) {
-                return@setOnClickListener
-            }
-
-            var beforeValue = unit.oneMLValue * amoutValue // 10(amount) L (unit)를
-            var changeValue = tool.oneMLValue // ml로 바꾸면
-
-            if (ingredient != null) {
-                var weight = ingredient.unitValue
-                beforeValue = beforeValue * weight
-            }
-
-            var result: Double = beforeValue / changeValue
-
-            if (humanOneValue > 1 || humanTwoValue > 1) {
-                result /= humanOneValue
-                result *= humanTwoValue
-
-            }
-
-            result = Math.round(result*100) / 100.0
-
-            var text = "$result ${tool.abbreviation}"
-            txt_calc_result.text = "$text 이다"
-
-            realm.beginTransaction()
-            val newItemId = newId()
-            val newItem = realm.createObject(CalcHistory::class.java, newItemId)
-            newItem.calcResultBefore = "$amoutValue${unit.abbreviation}"
-            newItem.calcResultAfter = text
-            newItem.createDate = Date().time
-            realm.commitTransaction()
-
-            btn_calc_button.visibility = View.GONE
-            layout_calc_result.visibility = View.VISIBLE
-            calc_frame_layout.visibility = View.INVISIBLE
+            unitRecyclerView.addItems(calculatorViewModel.convertUnitItemList())
         }
 
         layout_calc_delete_wrap.setOnClickListener {
             btn_calc_button.visibility = View.VISIBLE
             layout_calc_result.visibility = View.GONE
             calc_frame_layout.visibility = View.VISIBLE
-        }
-    }
-
-    private fun newId(): Long {
-        val maxId = realm.where(CalcHistory::class.java).max("historyId")
-        if (maxId != null) {
-            return maxId.toLong() + 1
-        }
-        return 0
-    }
-
-    private fun itemChange(containerCase: Int) {
-
-        Log.d("TAG", "item change case $containerCase")
-        if (ingredientSelectStatus[containerCase]) {
-            txt_ingredient.setTextColor(Color.parseColor("#131c32"))
-
-            val face: Typeface = ResourcesCompat.getFont(context!!, R.font.nanum_square_b)!!
-            txt_ingredient.typeface = face
-        } else {
-            txt_ingredient.setTextColor(Color.parseColor("#33131c32"))
-
-            val face: Typeface = ResourcesCompat.getFont(context!!, R.font.nanum_square_r)!!
-            txt_ingredient.typeface = face
-        }
-
-        if (portionSelectStatus[containerCase]) {
-            txt_human.setTextColor(Color.parseColor("#131c32"))
-
-            val face: Typeface = ResourcesCompat.getFont(context!!, R.font.nanum_square_b)!!
-            txt_human.typeface = face
-        } else {
-            txt_human.setTextColor(Color.parseColor("#33131c32"))
-
-            val face: Typeface = ResourcesCompat.getFont(context!!, R.font.nanum_square_r)!!
-            txt_human.typeface = face
-        }
-
-        if (plusSelectStatus[containerCase]) {
-            txt_calc_plus.setTextColor(Color.parseColor("#131c32"))
-
-            val face: Typeface = ResourcesCompat.getFont(context!!, R.font.nanum_square_b)!!
-            txt_calc_plus.typeface = face
-        } else {
-            txt_calc_plus.setTextColor(Color.parseColor("#33131c32"))
-
-            val face: Typeface = ResourcesCompat.getFont(context!!, R.font.nanum_square_r)!!
-            txt_calc_plus.typeface = face
-        }
-        updateContents(containerCase)
-    }
-
-    var humanTopWidth = 0
-    var humanBottomWidth = 0
-    var ingredientTopWidth = 0
-    var ingredientBottomWidth = 0
-
-    private fun updateContents(number: Int) {
-
-        var currentContainerCase = this.currentContainerCase
-        when (number) {
-            0 -> {
-                // ㅁ -> 재료
-                when (currentContainerCase) {
-                    0 -> {
-                        // 아무런 변화가 일어나지 않는다.
-                    }
-                    1 -> {
-                        Log.e("Error", "가능한 케이스가 아닙니다.")
-                        return
-                    }
-                    2 -> {
-                        humanTopWidth = hide_layout_human_top_standard.measuredWidth
-                        humanBottomWidth = hide_layout_human_bottom_standard.measuredWidth
-
-                        val animation = ValueAnimator.ofInt(layout_human_top_standard.measuredWidth, 0).setDuration(200)
-                        animation.addUpdateListener {
-                            val value = it.animatedValue as Int
-                            layout_human_top_standard.layoutParams.width = value
-                            layout_human_top_standard.requestLayout()
-                        }
-
-                        val animation2 =
-                            ValueAnimator.ofInt(layout_human_bottom_standard.measuredWidth, 0).setDuration(200)
-                        animation2.addUpdateListener {
-                            val value = it.animatedValue as Int
-                            layout_human_bottom_standard.layoutParams.width = value
-                            layout_human_bottom_standard.requestLayout()
-                        }
-
-                        val set = AnimatorSet()
-                        set.playSequentially(listOf(animation, animation2))
-                        set.interpolator = AccelerateDecelerateInterpolator()
-                        set.start()
-                    }
-
-                }
-
+            if (calculatorViewModel.currentCalcState.value == CalcTypeState.MATERIAL) {
+                calculatorViewModel.selectedEditBox.value = EditTextType.AMOUNT
+            } else {
+                calculatorViewModel.selectedEditBox.value = EditTextType.HUMAN_ONE
             }
-            1 -> {
-                // ㅁ -> 인원
-                when (currentContainerCase) {
-                    0 -> {
-                        Log.e("Error", "가능한 케이스가 아닙니다.")
-                        return
-                    }
-                    1 -> {
-                        // 아무런 변화가 일어나지 않는다.
-                    }
-                    2 -> {
-                        ingredientTopWidth = hide_layout_ingredient_top_standard.measuredWidth
-                        ingredientBottomWidth = hide_layout_ingredient_bottom_standard.measuredWidth
-                        val animation2 =
-                            ValueAnimator.ofInt(layout_ingredient_bottom_standard.measuredWidth, 0).setDuration(200)
-                        animation2.addUpdateListener {
-                            val value = it.animatedValue as Int
-                            layout_ingredient_bottom_standard.layoutParams.width = value
-                            layout_ingredient_bottom_standard.requestLayout()
-                        }
+        }
 
-                        val set = AnimatorSet()
-//                        set.playSequentially(listOf(animation, animation2))
-                        set.play(animation2)
-                        set.interpolator = AccelerateDecelerateInterpolator()
-                        set.start()
-                    }
-                }
+        val items: ArrayList<String> = calculatorViewModel.foodList.map { it -> it.unitNameBold } as ArrayList<String>
 
+        val picker = BaseWheelPicker(view, items, onChange)
+    }
+
+    private val onChange: (value: String) -> Unit = {
+        calculatorViewModel.onSelectIngredient(it)
+    }
+
+    private val changeItemList = Observer<CalcUnitType> {
+        unitRecyclerView.addItems(calculatorViewModel.convertUnitItemList())
+        when(calculatorViewModel.selectedUnitType.value) {
+            CalcUnitType.LIFE -> {
+                txt_unit_changer.text = "생활단위"
             }
-            2 -> {
-                // 재료 + 인원
-                when (currentContainerCase) {
-                    0 -> {
-                        val animation =
-                            ValueAnimator.ofInt(0, humanTopWidth).setDuration(200)
-                        animation.addUpdateListener {
-                            val value = it.animatedValue as Int
-                            layout_human_top_standard.layoutParams.width = value
-                            layout_human_top_standard.requestLayout()
-                        }
-
-                        val animation2 =
-                            ValueAnimator.ofInt(0, humanBottomWidth).setDuration(200)
-                        animation2.addUpdateListener {
-                            val value = it.animatedValue as Int
-                            layout_human_bottom_standard.layoutParams.width = value
-                            layout_human_bottom_standard.requestLayout()
-                        }
-
-
-                        val set = AnimatorSet()
-                        set.playSequentially(listOf(animation, animation2))
-                        set.interpolator = AccelerateDecelerateInterpolator()
-                        set.start()
-                    }
-
-                    1 -> {
-                        val animation2 =
-                            ValueAnimator.ofInt(0, ingredientBottomWidth).setDuration(200)
-                        animation2.addUpdateListener {
-                            val value = it.animatedValue as Int
-                            layout_ingredient_bottom_standard.layoutParams.width = value
-                            layout_ingredient_bottom_standard.requestLayout()
-                        }
-
-                        val set = AnimatorSet()
-                        set.play(animation2)
-                        set.interpolator = AccelerateDecelerateInterpolator()
-                        set.start()
-                    }
-
-                    2 -> {
-                        // 아무 일도 일어나지 않는다.
-                    }
-                }
+            CalcUnitType.NORMAL -> {
+                txt_unit_changer.text = "일반단위"
             }
             else -> {
-
+                Log.e("CalcFragment", "잘못된 타입입니다.")
             }
         }
-
-        this.currentContainerCase = number
-    }
-
-    fun showWeight() {
-
-        val weightWidth = hide_layout_ingredient_wrap.layoutParams.width
-
-        val animation =
-            ValueAnimator.ofInt(0, weightWidth).setDuration(200)
-        animation.addUpdateListener {
-            val value = it.animatedValue as Int
-            layout_ingredient_wrap.layoutParams.width = value
-            layout_ingredient_wrap.requestLayout()
-        }
-
-        val set = AnimatorSet()
-        set.play(animation)
-        set.interpolator = AccelerateDecelerateInterpolator()
-        set.start()
-    }
-
-    fun hideWeight() {
-
-        val weightWidth = hide_layout_ingredient_wrap.layoutParams.width
-
-        val animation =
-            ValueAnimator.ofInt(weightWidth, 0).setDuration(200)
-        animation.addUpdateListener {
-            val value = it.animatedValue as Int
-            layout_ingredient_wrap.layoutParams.width = value
-            layout_ingredient_wrap.requestLayout()
-        }
-
-        val set = AnimatorSet()
-        set.play(animation)
-        set.interpolator = AccelerateDecelerateInterpolator()
-        set.start()
-    }
-
-    private fun changeCalcContainerLayout(number: Int) {
-        val containerCase = calculatorViewModel.convertFragment(number)
-        itemChange(containerCase)
     }
 
     private fun callUnitSelector() {
         context!!.setTheme(R.style.ActionSheetStyleiOS7)
 
-        /* TODO: 생활단위, 일반단위를 고정값이 아니게 바꾸자 */
         ActionSheet.createBuilder(context, fragmentManager)
             .setCancelButtonTitle("취소")
             .setOtherButtonTitles("생활단위", "일반단위")
@@ -500,13 +289,9 @@ class CalcFragment : BaseFragment() {
             .setListener(object : ActionSheet.ActionSheetListener {
                 override fun onOtherButtonClick(actionSheet: ActionSheet?, index: Int) {
                     if (index == 0) {
-                        txt_unit_changer.text = "생활단위"
-                        calculatorViewModel.currentSelectedType = 1
-                        setRecyclerViewUnitModel(TYPE_LIFE)
+                        calculatorViewModel.selectedUnitType.value = CalcUnitType.LIFE
                     } else {
-                        txt_unit_changer.text = "일반단위"
-                        calculatorViewModel.currentSelectedType = 0
-                        setRecyclerViewUnitModel(TYPE_NORMAL)
+                        calculatorViewModel.selectedUnitType.value = CalcUnitType.NORMAL
                     }
                 }
 
@@ -516,76 +301,32 @@ class CalcFragment : BaseFragment() {
             }).show()
     }
 
-    private fun selectUnitItems(type: Int): ArrayList<UnitModel> {
-        val list: RealmResults<MeasureUnit> = when (type) {
-            TYPE_LIFE -> {
-                realm.where(MeasureUnit::class.java).equalTo("unitType", TYPE_LIFE).findAll()
-            }
-            TYPE_NORMAL -> {
-                realm.where(MeasureUnit::class.java).equalTo("unitType", TYPE_NORMAL).findAll()
-            }
-            else -> {
-                Log.e("TAG", "선언되지 않은 타입입니다.")
-                return arrayListOf()
-            }
-        }
+    private fun calculation() {
 
-        val result: ArrayList<UnitModel> = arrayListOf()
-
-        calculatorViewModel.itemSize = list.size
-
-        var index = calculatorViewModel.index
-        var size = 9
-        var max = (list.size / 10) + 1
-
-        if (index > max - 1) {
-            index = max - 1
-        }
-
-        var startIndex = index * size
-        var endIndex = index * size + size
-
-        if (endIndex > list.size) {
-            endIndex = list.size
-        }
-
-
-        for (x in startIndex until endIndex) {
-            result.add(
-                UnitModel(
-                    list[x].unitNameBold,
-                    list[x].unitNameSoft,
-                    list[x].unitType,
-                    list[x].isWeight,
-                    list[x].unitValue
-                )
-            )
-        }
-
-        return result
+        calculatorViewModel.selectedEditBox.value = EditTextType.NONE
+        txt_calc_result.text = calculatorViewModel.calculation()
     }
 
-    private fun setRecyclerViewUnitModel(type: Int) {
-        unitRecyclerView.removeAll()
-        unitRecyclerView.addItems(selectUnitItems(type))
+    private val adapterListener = object : AdapterListener {
+        override fun click(item: UnitModel) {
+            calculatorViewModel.onSelectUnit(item)
+        }
     }
+
 
     override fun subscribeUI() {
     }
 
     override fun onDataBinding(inflater: LayoutInflater, container: ViewGroup?): View? {
         binding = DataBindingUtil.inflate(inflater, R.layout.fragment_calc, container, false)
-//        binding.calculatorVM = calculatorViewModel
+        binding.vm = calculatorViewModel
         calculatorViewModel.context = context!!
         binding.lifecycleOwner = this
-
-        realm = Realm.getDefaultInstance()
 
         return binding.root
     }
 
     override fun onDestroy() {
-        calculatorViewModel.init()
         super.onDestroy()
     }
 }
