+++ app/src/main/res/layout/fragment_calc.xml
@@ -205,7 +205,7 @@
 
                 <include
                     android:id="@+id/calc_layout_ingredient"
-                    layout="@layout/layout_ingredient_scroll_wrap"
+                    layout="@layout/layout_wheel_picker"
                     android:layout_width="match_parent"
                     android:layout_height="match_parent"
                     android:visibility="gone" />
